プロンプトやニュースの取得方法を変更
