+++ src/lib/fetchNewsFromApi.ts
@@ -1,20 +1,4 @@
-interface NewsResponse {
-  status: string;
-  totalResults: number;
-  articles: Article[];
-}
-interface Article {
-  source: {
-    id: string | null;
-    name: string;
-  };
-  author: string | null;
-  title: string;
-  description: string;
-  url: string;
-  publishedAt: string;
-  content: string | null;
-}
+import { NewsResponse } from "@/interfaces/interface";
 
 const fetchNewsFromApi = async (): Promise<NewsResponse> => {
   const response = await fetch("/api/news");
